multichain-chainflip/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
motor==3.6.0
pymongo==4.9.0
web3==7.12.0
python-dotenv==1.0.1
cryptography==45.0.3
pydantic-settings==2.9.1
scikit-learn==1.6.1
requests==2.32.3
aiofiles==24.1.0
Pillow==11.0.0
qrcode==8.0
httpx==0.28.1
numpy==1.26.4
pandas==2.2.3
tensorflow==2.18.0
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1
aiohttp==3.11.10
websockets==14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
jinja2==3.1.4
itsdangerous==2.2.0
orjson==3.10.12
bcrypt==4.2.1
PyJWT==2.9.0